# swaggerdocs/swagger.yaml
info:
  contact:
    email: plazas@janelia.hhmi.org
    name: Stephen Plaza
  description: |-
    REST interface for neuPrint.  To test out the interface, copy  your token
    under your acocunt information. Then authorize Swagger by typing "Bearer " and
    pasting the token.
  title: neuprint API
  version: 0.1.0
paths:
...
  /logout:
    post:
      description: Clears session cookie for the user
      operationId: logoutHandler
      responses:
        "200":
          description: successful operation
      security:
      - Bearer: []
      summary: Logout user
      tags:
      - user
...
  /token:
    get:
      description: JWT token should be passed in header for authentication
      operationId: tokenHandler
      responses:
        "200":
          description: successful operation
      security:
      - Bearer: []
      summary: Returns JWT user bearer token
      tags:
      - user
security:
- Bearer: []
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    scopes:
      admin: Admin scope
      user: User scope
    type: apiKey
swagger: "2.0"